creator control on Activity added and working

+++ src/main/java/com/aelion/appliActivite/controllers/ActivityController.java
@@ -5,6 +5,8 @@ import javax.validation.Valid;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,8 +16,14 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.aelion.appliActivite.dto.ActivityFullDTO;
+import com.aelion.appliActivite.dto.UserFullDTO;
+import com.aelion.appliActivite.exceptions.NotAuthorizedException;
 import com.aelion.appliActivite.persistances.entities.Activity;
+import com.aelion.appliActivite.persistances.entities.User;
 import com.aelion.appliActivite.services.IActivityService;
+import com.aelion.appliActivite.services.IAuthChecker;
+import com.aelion.appliActivite.services.IUserService;
+import com.aelion.appliActivite.services.impl.AuthChecker;
 
 @RestController
 @RequestMapping(path = "/member/activity")
@@ -25,12 +33,26 @@ public class ActivityController {
 	IActivityService activityService;
 	
 	@Autowired
+	IUserService userService;
+	
+	@Autowired
+	IAuthChecker authChecker;
+	
+	
+	@Autowired
 	ModelMapper mapper;
 
 	
+	
+	public UserFullDTO getActivityCreator(Long id) {
+		
+		
+		return mapper.map(this.activityService.findOne(id).getCreator(), UserFullDTO.class);
+	}
 
 	@GetMapping("/{id}")
 	public ActivityFullDTO getActivityById(@PathVariable(name = "id") Long id) {
+		
 		return mapper.map(this.activityService.findOne(id), ActivityFullDTO.class);
 	}
 
@@ -41,7 +63,15 @@ public class ActivityController {
 
 	@DeleteMapping("/{id}")
 	public boolean deleteActivityById(@PathVariable(name = "id") Long id) {
-		return this.activityService.deleteById(id);
+		
+		if(this.authChecker.getCurrentUser().getId().equals(this.activityService.findOne(id).getCreator().getId())) {
+			return this.activityService.deleteById(id);
+			
+		}else {
+			throw new NotAuthorizedException("You don't have the necessary rights to delete this activity !");
+		}
+	
+		
 	}
